+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+version v0.2.0 [09/02/23]
+
+#feature
+
+- create pages screen and overview basic;
+- create pages context sidebar basic;
+
+---
+
 version v0.1.0 [09/02/23]
 
 #feature